group_remove, group_rename, major restructure
