+++ changelog.md
@@ -14,5 +14,11 @@
 ### Version 0.3.0
 - Updated selection macros according to Gromacs definitions, and introduced new macros for identifying DNA (`@dna`) and RNA (`@rna`).
 - Introduced the `System::group_create_ignore_warnings` function, which replicates the behavior of `System::group_create` while ignoring any WARNINGS that may occur during its execution. This new function addresses the issue of easily filtering out warnings returned by `System::group_create`, which uses a generic `Box<dyn Error>` return type. All other errors are propagated as usual since they indicate that the function failed.
-- Renamed functions `System::group_by_resid` and `System::group_by_resname` to `System::atoms_split_by_resid` and `System::atoms_split_by_resname`, respectively. **This is a BREAKING CHANGE.**
-- Introduced functions `System::group_split_by_resid` and `System::group_split_by_resname` for splitting groups of atoms by their residue ID and residue name, respectively.
+- **Breaking Change:** Renamed functions `System::group_by_resid` and `System::group_by_resname` to `System::atoms_split_by_resid` and `System::atoms_split_by_resname`, respectively.
+- Introduced new functions: `System::group_split_by_resid` for splitting groups of atoms by their residue ID and `System::group_split_by_resname` for splitting by residue name.
+- Added functions `System::group_remove` for removing groups and `System::group_rename` for renaming them.
+- Restructured codebase:
+  - Split the `system.rs` file into multiple smaller files.
+  - Grouped files related to IO operations into a dedicated directory.
+  - Organized files containing `System`-related methods into a separate directory.
+  - Placed files defining fundamental data structures into their own directory.
